Updates flake dependencies
Updates the flake.lock file to reflect the latest versions of dependencies.

This change includes updates to nix-darwin, flake-parts, flake-utils, ghostty, git-hooks.nix, hercules-ci-effects, home-manager, neovim-nightly-overlay, neovim, nixpkgs, and zig-overlay.

Also, adds flake-compat as a dependency for zig-overlay.

Enables zig overlay and removes yazi overlay and rust-overlay.

This ensures that the project is using the most up-to-date versions of its dependencies and resolves compatibility issues.